hackathon-1/Physical-AI-Humanoid-Robotic-Book/Physical-AI-Humanoid-Robotic-Book/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
cohere==4.9.0
openai>=1.3.5
psycopg2-binary==2.9.9
asyncpg==0.29.0
qdrant-client==1.8.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
sqlalchemy==2.0.23
alembic==1.12.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
